requirements.txt
djangorestframework==3.14.0
Django==4.2.2
psycopg2==2.9.6
gunicorn==20.1.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.14
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
drf-spectacular==0.26.2
pytest==7.3.2
pytest_factoryboy==2.5.1
pydantic==1.10.9
